working of the mac

+++ strokerenderer.cpp
@@ -1,7 +1,7 @@
 #include "strokerenderer.h"
 #include "timeline.h"
 #include <limits>
-
+#include <qmath.h>
 #include <vector>
 
 
@@ -128,7 +128,7 @@ void StrokeRenderer::addStrokeSprite(float x, float y)
 
     if(spriteCounter < N_SPRITES)
     {
-        GLushort posArray[] = {x * USHRT_MAX, y * USHRT_MAX};
+        GLushort posArray[] = {(GLushort)(x * USHRT_MAX), (GLushort)(y * USHRT_MAX)};
 
         glBindBuffer(GL_ARRAY_BUFFER, verticesId);
         glBufferSubData(GL_ARRAY_BUFFER, spriteCounter*VERTEX_COORD_SIZE, VERTEX_COORD_SIZE, posArray);
